Move to code and code cleanup
