+++ engine/src/main/java/org/apache/hop/core/truststore/CustomSocketFactory.java
@@ -19,24 +19,26 @@
  * limitations under the License.
  *
  ******************************************************************************/
-package org.apache.hop.pipeline.store;
+package org.apache.hop.core.truststore;
 
+import org.apache.hop.core.exception.HopException;
+import org.apache.hop.core.util.Utils;
+import org.apache.hop.i18n.BaseMessages;
+
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.TrustManager;
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.Socket;
 import java.security.KeyManagementException;
 import java.security.KeyStore;
 import java.security.NoSuchAlgorithmException;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSocketFactory;
-import javax.net.ssl.TrustManager;
-import org.apache.hop.core.exception.HopException;
-import org.apache.hop.core.util.Utils;
-import org.apache.hop.i18n.BaseMessages;
 
 public class CustomSocketFactory extends SSLSocketFactory {
 
-  private static final Class<?> PKG = CustomSocketFactory.class; // for i18n purposes, needed by Translator!!
+  private static final Class<?> classFromPackage =
+      CustomSocketFactory.class; // for i18n purposes, needed by Translator!!
 
   private static boolean configured;
 
@@ -89,7 +91,9 @@ public class CustomSocketFactory extends SSLSocketFactory {
       }
     } catch (Exception e) {
       throw new HopException(
-          BaseMessages.getString(PKG, "HopTrustManager.Exception.CouldNotCreateCertStore"), e);
+          BaseMessages.getString(
+              classFromPackage, "HopTrustManager.Exception.CouldNotCreateCertStore"),
+          e);
     }
 
     trustManagers = new HopTrustManager[] {new HopTrustManager(keyStore, path, password)};
