+++ engine/src/main/resources/org/apache/hop/core/truststore/messages/messages_en_US.properties
@@ -1,0 +1,4 @@
+HopTrustManager.Exception.CouldNotCreateCertStore=It was not possible to create key store
+HopTrustManager.Exception.CouldNotInitializeHopTrustManager=It was not possible to initialize the Hop trust manager
+HopTrustManager.Exception.CouldNotInitializeTrustManager=Could not initialize trust manager
+HopTrustManager.Exception.CouldNotOpenCertStore=It was not possible to open key store
