+++ engine/src/main/resources/org/apache/hop/core/truststore/messages/messages_fr_FR.properties
@@ -1,0 +1,3 @@
+HopTrustManager.Exception.CouldNotCreateCertStore=Impossible de cr\u00E9er le trousseau de cl\u00E9s.
+HopTrustManager.Exception.CouldNotOpenCertStore=Impossible d''ouvrir le trousseau de cl\u00E9
+HopTrustManager.Exception.CouldNotInitializeHopTrustManager=Impossible d''initialiser cr\u00E9er le trousseau de cl\u00E9s.
