+++ engine/src/main/resources/org/apache/hop/core/truststore/messages/messages_it_IT.properties
@@ -1,0 +1,4 @@
+HopTrustManager.Exception.CouldNotCreateCertStore=Impossibile creare il key store
+HopTrustManager.Exception.CouldNotInitializeTrustManager=Impossibile inizializzare il trust manager
+HopTrustManager.Exception.CouldNotOpenCertStore=Impossibile aprire il key store
+HopTrustManager.Exception.CouldNotInitializeHopTrustManager=Impossibile inizializzare il trust manager di Hop
