+++ plugins/transforms/ldap/src/main/resources/org/apache/hop/pipeline/transforms/ldapinput/messages/messages_en_US.properties
@@ -41,7 +41,6 @@ LdapInputDialog.FieldsTable.Format.Column=Format
 LdapInputDialog.FieldsTable.Attribute.Column=Attribute
 LdapInputDialog.setPaging.Label=Set paging
 LdapInputDialog.Group.CertificateGroup.Label=Certificate
-HopTrustManager.Exception.CouldNotCreateCertStore=It was not possible to create key store
 LdapInputMeta.SearchScope.OneLevel=One level scope
 LdapInputDialog.FilterString.Label=Filter String
 LdapInputField.TrimType.Both=both
@@ -56,7 +55,6 @@ LdapInputDialog.MultiValuedSeparator.Tooltip=If several values are returned for 
 LdapConnection.Error.Update=Error updating entry with dn \= [{0}]\!
 LdapInputDialog.TrustStorePath.Tooltip=Trust store path
 LdapInput.SchemaList.Message=The schema \: \n{0}\n
-HopTrustManager.Exception.CouldNotInitializeTrustManager=Could not initialize trust manager
 LdapInputDialog.FieldsTable.Group.Column=Group
 LdapInputDialog.Search.Tab=Search
 LdapInputField.TrimType.None=none
@@ -133,7 +131,6 @@ LdapInputDialog.Fields.Tab=Fields
 LdapInputDialog.SearchBase.Tooltip=Search base
 LdapInput.ClassUsed.Message=The class used to connect is [{0}]
 LdapInputDialog.Connected.NOK=Not connected to LDAP server\!
-HopTrustManager.Exception.CouldNotOpenCertStore=It was not possible to open key store
 LdapInputDialog.ErrorPreviewingData.DialogMessage=An error occurred while trying to preview data
 LdapConnection.Update.Attribute=Will update attribute [{0}] with value [{1}]...
 LdapConnection.Error.RetrievingFields=Error retrieving fields\!
@@ -155,4 +152,3 @@ LdapInput.log.Disconnection.Done=Disconnected from server
 LdapConnection.Error.Search=Error while performing search\!
 LdapInputField.TrimType.Right=right
 LdapInputDialog.FailedToGetFields.DialogMessage=Unable to get fields from previous transforms
-HopTrustManager.Exception.CouldNotInitializeHopTrustManager=It was not possible to initialize the Hop trust manager
