+++ plugins/transforms/ldap/src/main/resources/org/apache/hop/pipeline/transforms/ldapinput/messages/messages_fr_FR.properties
@@ -39,7 +39,6 @@ LdapInputDialog.FieldsTable.Format.Column=Format
 LdapInputDialog.FieldsTable.Attribute.Column=Attribut
 LdapInputDialog.setPaging.Label=Activer mode page
 LdapInputDialog.Group.CertificateGroup.Label=Certificat
-HopTrustManager.Exception.CouldNotCreateCertStore=Impossible de cr\u00E9er le trousseau de cl\u00E9s.
 LdapInputMeta.SearchScope.OneLevel=Rechercher sur un niveau de l''objet
 LdapInputDialog.FilterString.Label=Chaine filtre
 LdapInputField.TrimType.Both=les deux
@@ -129,7 +128,6 @@ LdapInputDialog.Fields.Tab=Champs
 LdapInputDialog.SearchBase.Tooltip=Base recherche
 LdapInput.ClassUsed.Message=La class utilis\u00E9e est [{0}]
 LdapInputDialog.Connected.NOK=Non connect\u00E9 au serveur LDAP\!\n Exception \: \n[{0}]
-HopTrustManager.Exception.CouldNotOpenCertStore=Impossible d''ouvrir le trousseau de cl\u00E9
 LdapInputDialog.ErrorPreviewingData.DialogMessage=PDI a rencontr\u00E9 une erreur lors de la tentative de pr\u00E9visualisation
 LdapConnection.Update.Attribute=mise \u00E0 jour de l''attribut [{0}] avec la valeur [{1}]...
 LdapConnection.Error.RetrievingFields=Erreur lors de la r\u00E9cup\u00E9ration des champs\!
@@ -150,4 +148,3 @@ LdapInputDialog.FieldsTable.Name.Column=Nom
 LdapInput.log.Disconnection.Done=D\u00E9connect\u00E9 du serveur
 LdapInputField.TrimType.Right=droite
 LdapInputDialog.FailedToGetFields.DialogMessage=Impossible de r\u00E9cup\u00E9rer les champs depuis les \u00E9tapes pr\u00E9c\u00E9dentes\!
-HopTrustManager.Exception.CouldNotInitializeHopTrustManager=Impossible d''initialiser cr\u00E9er le trousseau de cl\u00E9s.
