+++ plugins/transforms/ldap/src/main/resources/org/apache/hop/pipeline/transforms/ldapinput/messages/messages_it_IT.properties
@@ -35,7 +35,6 @@ LdapInputDialog.FieldsTable.Format.Column=Formato
 LdapInputDialog.FieldsTable.Attribute.Column=Attributo
 LdapInputDialog.setPaging.Label=Imposta paginazione
 LdapInputDialog.Group.CertificateGroup.Label=Certificato
-HopTrustManager.Exception.CouldNotCreateCertStore=Impossibile creare il key store
 LdapInputMeta.SearchScope.OneLevel=Scope a livello 1
 LdapInputDialog.FilterString.Label=Stringa di filtro
 LdapInputField.TrimType.Both=entrambi
@@ -50,7 +49,6 @@ LdapInputDialog.MultiValuedSeparator.Tooltip=Se diversi valori vengono ritornati
 LdapConnection.Error.Update=Errore nell''aggiornamento dell''entry con dn \= [{0}]\!
 LdapInputDialog.TrustStorePath.Tooltip=Repository del certificato
 LdapInput.SchemaList.Message=Lo schema\:\n{0}\n
-HopTrustManager.Exception.CouldNotInitializeTrustManager=Impossibile inizializzare il trust manager
 LdapInputDialog.FieldsTable.Group.Column=Gruppo
 LdapInputDialog.Search.Tab=Ricerca
 LdapInputField.TrimType.None=nessuna
@@ -123,7 +121,6 @@ LdapInputDialog.Fields.Tab=Campi
 LdapInputDialog.SearchBase.Tooltip=Base di ricerca
 LdapInput.ClassUsed.Message=La classe usata per connettersi \u00E8 [{0}]
 LdapInputDialog.Connected.NOK=Non connesso al server LDAP\!\nEccezione\:\n[{0}]
-HopTrustManager.Exception.CouldNotOpenCertStore=Impossibile aprire il key store
 LdapInputDialog.ErrorPreviewingData.DialogMessage=Errore durante l''anteprima dei dati
 LdapConnection.Update.Attribute=Verr\u00E0 aggiornato l''attributo [{0}] con il valore [{1}]...
 LdapConnection.Error.RetrievingFields=Errore nel recupero dei campi\!
@@ -144,4 +141,3 @@ LdapInputDialog.FieldsTable.Name.Column=Nome
 LdapInput.log.Disconnection.Done=Disconnesso dal server
 LdapInputField.TrimType.Right=destra
 LdapInputDialog.FailedToGetFields.DialogMessage=Impossibile ricevere i campi dai passi precedenti
-HopTrustManager.Exception.CouldNotInitializeHopTrustManager=Impossibile inizializzare il trust manager di Hop
